# Improve perf upwind