--- a/kernels/vlasov/vlasov_boundary_surfvx_1x1v_ser_p1.c
+++ b/kernels/vlasov/vlasov_boundary_surfvx_1x1v_ser_p1.c
@@ -1,4 +1,5 @@
 #include <gkyl_vlasov_kernels.h> 
+#include <gkyl_basis_ser_1x1v_p1_surfvx_quad.h> 
 GKYL_CU_DH void vlasov_boundary_surfvx_1x1v_ser_p1(const double *w, const double *dxv, const double *qmem, const int edge, const double *fEdge, const double *fSkin, double* GKYL_RESTRICT out) 
 { 
   // w:           Cell-center coordinates.
@@ -11,37 +12,37 @@ GKYL_CU_DH void vlasov_boundary_surfvx_1x1v_ser_p1(const double *w, const double
   const double dv1 = dxv[1], wv1 = w[1]; 
   const double *E0 = &qmem[0]; 
 
-  double Ghat[2]; 
-  double alpha[2]; 
+  double alpha[2] = {0.0}; 
 
   alpha[0] = E0[0]; 
   alpha[1] = E0[1]; 
 
-  double fUpwindQuad[2];
-  double fUpwind[2];
+  double fUpwindQuad[2] = {0.0};
+  double fUpwind[2] = {0.0};
+  double Ghat[2] = {0.0}; 
 
   if (edge == -1) { 
 
   if (alpha[0]-alpha[1] > 0) { 
 
-    fUpwindQuad[0] = (-0.8660254037844386*fSkin[3])+0.8660254037844386*fSkin[2]-0.5*fSkin[1]+0.5*fSkin[0]; 
+    fUpwindQuad[0] = ser_1x1v_p1_surfvx_quad_0(1, fSkin); 
   } else { 
 
-    fUpwindQuad[0] = 0.8660254037844386*fEdge[3]-0.8660254037844386*fEdge[2]-0.5*fEdge[1]+0.5*fEdge[0]; 
+    fUpwindQuad[0] = ser_1x1v_p1_surfvx_quad_0(-1, fEdge); 
   } 
   if (alpha[1]+alpha[0] > 0) { 
 
-    fUpwindQuad[1] = 0.8660254037844386*(fSkin[3]+fSkin[2])+0.5*(fSkin[1]+fSkin[0]); 
+    fUpwindQuad[1] = ser_1x1v_p1_surfvx_quad_1(1, fSkin); 
   } else { 
 
-    fUpwindQuad[1] = 0.5*(fEdge[1]+fEdge[0])-0.8660254037844386*(fEdge[3]+fEdge[2]); 
+    fUpwindQuad[1] = ser_1x1v_p1_surfvx_quad_1(-1, fEdge); 
   } 
 
   fUpwind[0] = 0.7071067811865475*(fUpwindQuad[1]+fUpwindQuad[0]); 
   fUpwind[1] = 0.7071067811865475*(fUpwindQuad[1]-1.0*fUpwindQuad[0]); 
 
-  Ghat[0] = 0.7071067811865475*(alpha[1]*fUpwind[1]+alpha[0]*fUpwind[0]); 
-  Ghat[1] = 0.7071067811865475*(alpha[0]*fUpwind[1]+fUpwind[0]*alpha[1]); 
+  Ghat[0] += 0.7071067811865475*(alpha[1]*fUpwind[1]+alpha[0]*fUpwind[0]); 
+  Ghat[1] += 0.7071067811865475*(alpha[0]*fUpwind[1]+fUpwind[0]*alpha[1]); 
 
   out[0] += -0.7071067811865475*Ghat[0]*dv10; 
   out[1] += -0.7071067811865475*Ghat[1]*dv10; 
@@ -52,24 +53,24 @@ GKYL_CU_DH void vlasov_boundary_surfvx_1x1v_ser_p1(const double *w, const double
 
   if (alpha[0]-alpha[1] > 0) { 
 
-    fUpwindQuad[0] = (-0.8660254037844386*fEdge[3])+0.8660254037844386*fEdge[2]-0.5*fEdge[1]+0.5*fEdge[0]; 
+    fUpwindQuad[0] = ser_1x1v_p1_surfvx_quad_0(1, fEdge); 
   } else { 
 
-    fUpwindQuad[0] = 0.8660254037844386*fSkin[3]-0.8660254037844386*fSkin[2]-0.5*fSkin[1]+0.5*fSkin[0]; 
+    fUpwindQuad[0] = ser_1x1v_p1_surfvx_quad_0(-1, fSkin); 
   } 
   if (alpha[1]+alpha[0] > 0) { 
 
-    fUpwindQuad[1] = 0.8660254037844386*(fEdge[3]+fEdge[2])+0.5*(fEdge[1]+fEdge[0]); 
+    fUpwindQuad[1] = ser_1x1v_p1_surfvx_quad_1(1, fEdge); 
   } else { 
 
-    fUpwindQuad[1] = 0.5*(fSkin[1]+fSkin[0])-0.8660254037844386*(fSkin[3]+fSkin[2]); 
+    fUpwindQuad[1] = ser_1x1v_p1_surfvx_quad_1(-1, fSkin); 
   } 
 
   fUpwind[0] = 0.7071067811865475*(fUpwindQuad[1]+fUpwindQuad[0]); 
   fUpwind[1] = 0.7071067811865475*(fUpwindQuad[1]-1.0*fUpwindQuad[0]); 
 
-  Ghat[0] = 0.7071067811865475*(alpha[1]*fUpwind[1]+alpha[0]*fUpwind[0]); 
-  Ghat[1] = 0.7071067811865475*(alpha[0]*fUpwind[1]+fUpwind[0]*alpha[1]); 
+  Ghat[0] += 0.7071067811865475*(alpha[1]*fUpwind[1]+alpha[0]*fUpwind[0]); 
+  Ghat[1] += 0.7071067811865475*(alpha[0]*fUpwind[1]+fUpwind[0]*alpha[1]); 
 
   out[0] += 0.7071067811865475*Ghat[0]*dv10; 
   out[1] += 0.7071067811865475*Ghat[1]*dv10;