--- a/kernels/vlasov/vlasov_surfvx_1x1v_ser_p1.c
+++ b/kernels/vlasov/vlasov_surfvx_1x1v_ser_p1.c
@@ -1,4 +1,5 @@
 #include <gkyl_vlasov_kernels.h> 
+#include <gkyl_basis_ser_1x1v_p1_surfvx_quad.h> 
 GKYL_CU_DH void vlasov_surfvx_1x1v_ser_p1(const double *w, const double *dxv, const double *qmem, const double *fl, const double *fc, const double *fr, double* GKYL_RESTRICT out) 
 { 
   // w:         Cell-center coordinates.
@@ -10,46 +11,47 @@ GKYL_CU_DH void vlasov_surfvx_1x1v_ser_p1(const double *w, const double *dxv, co
   const double dv1 = dxv[1], wv1 = w[1]; 
   const double *E0 = &qmem[0]; 
 
-  double alpha[2]; 
+  double alpha[2] = {0.0}; 
 
   alpha[0] = E0[0]; 
   alpha[1] = E0[1]; 
 
-  double fUpwindQuad_l[2];
-  double fUpwindQuad_r[2];
+  double fUpwindQuad_l[2] = {0.0};
+  double fUpwindQuad_r[2] = {0.0};
+  double fUpwind_l[2] = {0.0};;
+  double fUpwind_r[2] = {0.0};
+  double Ghat_l[2] = {0.0}; 
+  double Ghat_r[2] = {0.0}; 
+
   if (alpha[0]-alpha[1] > 0) { 
 
-    fUpwindQuad_l[0] = (-0.8660254037844386*fl[3])+0.8660254037844386*fl[2]-0.5*fl[1]+0.5*fl[0]; 
-    fUpwindQuad_r[0] = (-0.8660254037844386*fc[3])+0.8660254037844386*fc[2]-0.5*fc[1]+0.5*fc[0]; 
+    fUpwindQuad_l[0] = ser_1x1v_p1_surfvx_quad_0(1, fl); 
+    fUpwindQuad_r[0] = ser_1x1v_p1_surfvx_quad_0(1, fc); 
   } else { 
 
-    fUpwindQuad_l[0] = 0.8660254037844386*fc[3]-0.8660254037844386*fc[2]-0.5*fc[1]+0.5*fc[0]; 
-    fUpwindQuad_r[0] = 0.8660254037844386*fr[3]-0.8660254037844386*fr[2]-0.5*fr[1]+0.5*fr[0]; 
+    fUpwindQuad_l[0] = ser_1x1v_p1_surfvx_quad_0(-1, fc); 
+    fUpwindQuad_r[0] = ser_1x1v_p1_surfvx_quad_0(-1, fr); 
   } 
   if (alpha[1]+alpha[0] > 0) { 
 
-    fUpwindQuad_l[1] = 0.8660254037844386*(fl[3]+fl[2])+0.5*(fl[1]+fl[0]); 
-    fUpwindQuad_r[1] = 0.8660254037844386*(fc[3]+fc[2])+0.5*(fc[1]+fc[0]); 
+    fUpwindQuad_l[1] = ser_1x1v_p1_surfvx_quad_1(1, fl); 
+    fUpwindQuad_r[1] = ser_1x1v_p1_surfvx_quad_1(1, fc); 
   } else { 
 
-    fUpwindQuad_l[1] = 0.5*(fc[1]+fc[0])-0.8660254037844386*(fc[3]+fc[2]); 
-    fUpwindQuad_r[1] = 0.5*(fr[1]+fr[0])-0.8660254037844386*(fr[3]+fr[2]); 
+    fUpwindQuad_l[1] = ser_1x1v_p1_surfvx_quad_1(-1, fc); 
+    fUpwindQuad_r[1] = ser_1x1v_p1_surfvx_quad_1(-1, fr); 
   } 
-  double fUpwind_l[2];
   fUpwind_l[0] = 0.7071067811865475*(fUpwindQuad_l[1]+fUpwindQuad_l[0]); 
   fUpwind_l[1] = 0.7071067811865475*(fUpwindQuad_l[1]-1.0*fUpwindQuad_l[0]); 
 
-  double fUpwind_r[2];
   fUpwind_r[0] = 0.7071067811865475*(fUpwindQuad_r[1]+fUpwindQuad_r[0]); 
   fUpwind_r[1] = 0.7071067811865475*(fUpwindQuad_r[1]-1.0*fUpwindQuad_r[0]); 
 
-  double Ghat_l[2]; 
-  double Ghat_r[2]; 
-  Ghat_l[0] = 0.7071067811865475*(alpha[1]*fUpwind_l[1]+alpha[0]*fUpwind_l[0]); 
-  Ghat_l[1] = 0.7071067811865475*(alpha[0]*fUpwind_l[1]+fUpwind_l[0]*alpha[1]); 
+  Ghat_l[0] += 0.7071067811865475*(alpha[1]*fUpwind_l[1]+alpha[0]*fUpwind_l[0]); 
+  Ghat_l[1] += 0.7071067811865475*(alpha[0]*fUpwind_l[1]+fUpwind_l[0]*alpha[1]); 
 
-  Ghat_r[0] = 0.7071067811865475*(alpha[1]*fUpwind_r[1]+alpha[0]*fUpwind_r[0]); 
-  Ghat_r[1] = 0.7071067811865475*(alpha[0]*fUpwind_r[1]+fUpwind_r[0]*alpha[1]); 
+  Ghat_r[0] += 0.7071067811865475*(alpha[1]*fUpwind_r[1]+alpha[0]*fUpwind_r[0]); 
+  Ghat_r[1] += 0.7071067811865475*(alpha[0]*fUpwind_r[1]+fUpwind_r[0]*alpha[1]); 
 
   out[0] += (0.7071067811865475*Ghat_l[0]-0.7071067811865475*Ghat_r[0])*dv10; 
   out[1] += (0.7071067811865475*Ghat_l[1]-0.7071067811865475*Ghat_r[1])*dv10;