This commit implements a re-organization of the upwind fluxes to now use a suite of header functions in the basis/ directory which interpolate the function of interest onto surface quadrature points. We can now these header functions for other equations going forward. It may be worth consolidating some of them since they are quite small, but for now I just made them separate by both dimensionality and polynomial order.

In addition this commit implements a generalization of the upwind fluxes which eliminates aliasing for non-tensor bases. We now *always* evaluate the upwinding on a tensor product quadrature basis, regardless of the underlying basis, project that tensor product nodal basis back onto the modal basis, and evaluate Ghat = alpha * f_hat where alpha is the flux and f_hat is the upwinded f (the function being upwinded).

For tensor product modal bases (and polyOrder = 1 Serendipity) this projection from nodal to modal is "trivial," in that there is a one-to-one transformation which does not require the explicit evaluation of an inner product. For modal Serendipity, p>1, we evaluate the inner product of the modal surface basis functions and the nodal tensor product surface quadrature basis. 

This is a slight cost increase, but actually seems negligible, at least in the 4d Weibel problem. On my iMac the timings went from 

> Number of update calls 136
> Number of forward-Euler calls 408
> Number of RK stage-2 failures 0
> Number of RK stage-3 failures 0
> Species RHS calc took 16.1107 secs
> Field RHS calc took 0.010645 secs
> Current evaluation and accumulate took 0.199343 secs
> Updates took 17.6042 secs

To

> Number of update calls 136
> Number of forward-Euler calls 408
> Number of RK stage-2 failures 0
> Number of RK stage-3 failures 0
> Species RHS calc took 17.6609 secs
> Field RHS calc took 0.010388 secs
> Current evaluation and accumulate took 0.169811 secs
> Updates took 18.9457 secs

A similar 10 percent slow down is visible on GPUs, but this is an acceptable cost to me for peace of mind on this evaluation and projection step for p>1 Serendipity bases.